+++ src/pages/SignUp/components/Step2/index.tsx
